python/requirements.txt
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.3
Flask-SocketIO==5.3.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyasn1==0.4.8
pycparser==2.21
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
rsa==4.9
simple-websocket==0.9.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.14.0
